changed an iterator in for loop in countdown()
Previously it was hardly set iterator starting from 10. So, when video was skipped or paused at last 10 seconds of video and later, countdown was starting anyway at 10.
I've passed a refreshed timer, global variable, that is calculated from video length and its current time.
Countdown now is displaying real time to the end of video.
Details and explanation of that change can be seen within the code itself.